requirements.txt
beautifulsoup4==4.8.0
lxml==4.4.1
Pillow==6.1.0
selenium==3.141.0
soupsieve==1.9.3
urllib3==1.25.3
